backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Migrations
alembic==1.13.1

# Utils
python-dotenv==1.0.1
loguru==0.7.2
